+++ src/main/java/com/example/gabojago_server/dto/request/member/MemberRequestDto.java
@@ -1,9 +1,8 @@
-package com.example.gabojago_server.dto;
+package com.example.gabojago_server.dto.request.member;
 
-import com.example.gabojago_server.model.Authority;
-import com.example.gabojago_server.model.Member;
+import com.example.gabojago_server.model.member.Authority;
+import com.example.gabojago_server.model.member.Member;
 import lombok.*;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.crypto.password.PasswordEncoder;
