+++ src/main/java/com/example/gabojago_server/model/member/Authority.java
@@ -1,4 +1,4 @@
-package com.example.gabojago_server.model;
+package com.example.gabojago_server.model.member;
 
 public enum Authority {
     ROLE_USER, ROLE_ADMIN
