+++ src/main/java/com/example/gabojago_server/service/member/MemberService.java
@@ -1,9 +1,9 @@
-package com.example.gabojago_server.service;
+package com.example.gabojago_server.service.member;
 
 import com.example.gabojago_server.config.SecurityUtil;
-import com.example.gabojago_server.dto.MemberResponseDto;
-import com.example.gabojago_server.model.Member;
-import com.example.gabojago_server.repository.MemberRepository;
+import com.example.gabojago_server.dto.response.member.MemberResponseDto;
+import com.example.gabojago_server.model.member.Member;
+import com.example.gabojago_server.repository.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
